refactor(api): PR6 — extract thin wrappers (kill_switch + tune + health + notifications)
Move 17 routes across 4 domains to dedicated APIRouter modules:
- api/kill_switch.py (6 routes): recalibrate, recommendations CRUD, decisions, current_state
- api/tune.py (3 routes): latest, apply, reject
- api/health.py (5 routes): /health, /health/symbols, /health/events, /health/dashboard, /health/reactivate/{symbol}
- api/notifications.py (3 routes): /notifications, /notifications/{id}/read, /notifications/read-all

Each module is a thin router wrapper — most logic delegates to existing
strategy/kill_switch_v2*, health.py, db tables, and notifier system. Light
parity tests per domain (1-2 endpoints each) backed by snapshot baselines.

btc_api.py shrinks ~600 LOC. Routes still inline in btc_api.py: /, /symbols,
/status, /scan, /webhook/test (these are bootstrap-level / cross-cutting and
move in PR7 alongside the scanner runtime).

Fix existing calibrator tests that patched btc_api.load_config / save_config
directly — now also patch api.config.load_config and api.kill_switch.save_config
since the routes live outside btc_api.py after PR6.

1029 tests pass.
