+++ btc_api.py
@@ -85,6 +85,11 @@ from db.signals import (  # noqa: F401
     save_scan, get_scans, get_latest_signal, get_latest_scan, get_signals_summary,
 )
 from api.signals import router as signals_router
+# PR6: thin route wrappers extracted to dedicated APIRouter modules.
+from api.kill_switch import router as kill_switch_router
+from api.tune import router as tune_router
+from api.health import router as health_router
+from api.notifications import router as notifications_router
 
 
 # ─────────────────────────────────────────────────────────────────────────────
@@ -497,6 +502,10 @@ app.include_router(ohlcv_router)
 app.include_router(config_router)
 app.include_router(positions_router)
 app.include_router(signals_router)
+app.include_router(kill_switch_router)
+app.include_router(tune_router)
+app.include_router(health_router)
+app.include_router(notifications_router)
 
 
 # ── Endpoints ────────────────────────────────────────────────────────────────
@@ -570,296 +579,6 @@ def force_scan(
     return {"scanned": len(results), "results": results}
 
 
-@app.post(
-    "/kill_switch/recalibrate",
-    summary="Manually trigger an auto-calibrator recommendation",
-    dependencies=[Depends(verify_api_key)],
-)
-def kill_switch_recalibrate():
-    """Manually trigger the auto-calibrator (#187 B4b.1).
-
-    In B4b.1 this returns no_feasible (stub fitness). B4b.2 (#216) will
-    replace the stub with v2-aware backtest grid optimization.
-    """
-    from strategy.kill_switch_v2_calibrator import (
-        run_optimization_stub, _persist_recommendation,
-    )
-    from strategy.kill_switch_v2_optimizer import run_optimization_v2
-    from datetime import datetime, timezone
-
-    try:
-        cfg = load_config()
-        try:
-            result = run_optimization_v2(cfg)
-        except Exception as opt_err:
-            log.warning(
-                "run_optimization_v2 failed; falling back to stub: %s",
-                opt_err, exc_info=True,
-            )
-            result = run_optimization_stub(cfg)
-        now = datetime.now(tz=timezone.utc)
-        rec_id = _persist_recommendation(
-            triggered_by=["manual"], result=result, now=now,
-        )
-    except Exception as e:
-        log.error(
-            "POST /kill_switch/recalibrate failed: %s", e, exc_info=True,
-        )
-        raise HTTPException(
-            status_code=500,
-            detail=f"recalibrate failed: {type(e).__name__}: {e}",
-        )
-
-    log.warning(
-        "Kill switch v2: recomendación id=%d (status=%s, triggered_by=manual). "
-        "Telegram pendiente B4b.3.",
-        rec_id, result["status"],
-    )
-    return {"recommendation_id": rec_id, "status": result["status"]}
-
-
-@app.get(
-    "/kill_switch/recommendations",
-    summary="List auto-calibrator recommendations",
-    dependencies=[Depends(verify_api_key)],
-)
-def kill_switch_list_recommendations(
-    since: Optional[str] = Query(
-        None, description="ISO timestamp; only rows with ts >= since",
-    ),
-    status: Optional[str] = Query(
-        None,
-        description="Filter by status (pending, applied, ignored, "
-                    "superseded, no_feasible)",
-    ),
-    limit: int = Query(100, ge=1, le=1000),
-):
-    """List auto-calibrator recommendations, latest first."""
-    import json as _json
-
-    conn = get_db()
-    try:
-        sql = (
-            "SELECT id, ts, triggered_by, slider_value, projected_pnl, "
-            "projected_dd, status, applied_ts, applied_by, report_json "
-            "FROM kill_switch_recommendations WHERE 1=1"
-        )
-        params: list = []
-        if since:
-            sql += " AND ts >= ?"
-            params.append(since)
-        if status:
-            sql += " AND status = ?"
-            params.append(status)
-        sql += " ORDER BY ts DESC LIMIT ?"
-        params.append(int(limit))
-        rows = conn.execute(sql, params).fetchall()
-    finally:
-        conn.close()
-
-    result = []
-    for r in rows:
-        d = {
-            "id": r[0],
-            "ts": r[1],
-            "triggered_by": r[2],
-            "slider_value": r[3],
-            "projected_pnl": r[4],
-            "projected_dd": r[5],
-            "status": r[6],
-            "applied_ts": r[7],
-            "applied_by": r[8],
-            "report": None,
-        }
-        try:
-            d["triggered_by"] = _json.loads(d["triggered_by"])
-        except (TypeError, ValueError) as e:
-            log.warning(
-                "Corrupted recommendation row id=%s: triggered_by JSON "
-                "parse failed (%s); raw=%r", r[0], e, r[2],
-            )
-        try:
-            d["report"] = _json.loads(r[9])
-        except (TypeError, ValueError) as e:
-            log.warning(
-                "Corrupted recommendation row id=%s: report_json parse "
-                "failed (%s); raw=%r", r[0], e, r[9],
-            )
-            d["report"] = None
-        result.append(d)
-    return result
-
-
-@app.post(
-    "/kill_switch/recommendations/{rec_id}/apply",
-    summary="Apply a pending recommendation (operator action)",
-    dependencies=[Depends(verify_api_key)],
-)
-def kill_switch_apply_recommendation(rec_id: int):
-    """Apply a pending recommendation: write config override + mark applied.
-
-    Returns the updated row. 404 if not found, 400 if not pending.
-    """
-    from datetime import datetime, timezone
-
-    try:
-        conn = get_db()
-        try:
-            row = conn.execute(
-                """SELECT id, status, slider_value
-                   FROM kill_switch_recommendations WHERE id = ?""",
-                (rec_id,),
-            ).fetchone()
-        finally:
-            conn.close()
-
-        if row is None:
-            raise HTTPException(status_code=404, detail=f"recommendation {rec_id} not found")
-        if row[1] != "pending":
-            raise HTTPException(
-                status_code=400,
-                detail=f"recommendation {rec_id} is already {row[1]}",
-            )
-        slider_value = row[2]
-        if slider_value is None:
-            raise HTTPException(
-                status_code=400,
-                detail=f"recommendation {rec_id} has no slider_value to apply",
-            )
-        slider_int = int(slider_value)
-        if not (0 <= slider_int <= 100):
-            raise HTTPException(
-                status_code=400,
-                detail=(
-                    f"recommendation {rec_id} has out-of-range slider_value="
-                    f"{slider_int} (must be 0..100)"
-                ),
-            )
-
-        # Write config override. save_config only merges at kill_switch level
-        # (it replaces the entire v2 sub-dict), so do read-modify-write here to
-        # preserve other v2 keys (auto_calibrator, regime_adjustments, etc).
-        existing_cfg = load_config()
-        existing_v2 = (existing_cfg.get("kill_switch", {}) or {}).get("v2", {}) or {}
-        merged_v2 = {**existing_v2, "aggressiveness": slider_int}
-        save_config({"kill_switch": {"v2": merged_v2}})
-
-        # Update DB row
-        now_iso = datetime.now(tz=timezone.utc).isoformat()
-        conn = get_db()
-        try:
-            conn.execute(
-                """UPDATE kill_switch_recommendations
-                   SET status = 'applied', applied_ts = ?, applied_by = 'operator'
-                   WHERE id = ?""",
-                (now_iso, rec_id),
-            )
-            conn.commit()
-            updated = conn.execute(
-                """SELECT id, ts, triggered_by, slider_value, projected_pnl,
-                          projected_dd, status, applied_ts, applied_by, report_json
-                   FROM kill_switch_recommendations WHERE id = ?""",
-                (rec_id,),
-            ).fetchone()
-        finally:
-            conn.close()
-
-        log.warning(
-            "Kill switch v2: operator applied recomendación id=%d slider=%d",
-            rec_id, int(slider_value),
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        log.error(
-            "POST /kill_switch/recommendations/%d/apply failed: %s",
-            rec_id, e, exc_info=True,
-        )
-        raise HTTPException(
-            status_code=500,
-            detail=f"apply failed: {type(e).__name__}: {e}",
-        )
-
-    return {
-        "id": updated[0], "ts": updated[1], "triggered_by": updated[2],
-        "slider_value": updated[3], "projected_pnl": updated[4],
-        "projected_dd": updated[5], "status": updated[6],
-        "applied_ts": updated[7], "applied_by": updated[8],
-    }
-
-
-@app.post(
-    "/kill_switch/recommendations/{rec_id}/ignore",
-    summary="Ignore a pending recommendation (operator action)",
-    dependencies=[Depends(verify_api_key)],
-)
-def kill_switch_ignore_recommendation(rec_id: int):
-    """Mark a pending recommendation as ignored (no config change).
-
-    404 if not found, 400 if not pending.
-    """
-    from datetime import datetime, timezone
-
-    try:
-        conn = get_db()
-        try:
-            row = conn.execute(
-                """SELECT id, status FROM kill_switch_recommendations WHERE id = ?""",
-                (rec_id,),
-            ).fetchone()
-        finally:
-            conn.close()
-
-        if row is None:
-            raise HTTPException(status_code=404, detail=f"recommendation {rec_id} not found")
-        if row[1] != "pending":
-            raise HTTPException(
-                status_code=400,
-                detail=f"recommendation {rec_id} is already {row[1]}",
-            )
-
-        now_iso = datetime.now(tz=timezone.utc).isoformat()
-        conn = get_db()
-        try:
-            conn.execute(
-                """UPDATE kill_switch_recommendations
-                   SET status = 'ignored', applied_ts = ?, applied_by = 'operator'
-                   WHERE id = ?""",
-                (now_iso, rec_id),
-            )
-            conn.commit()
-            updated = conn.execute(
-                """SELECT id, ts, triggered_by, slider_value, projected_pnl,
-                          projected_dd, status, applied_ts, applied_by, report_json
-                   FROM kill_switch_recommendations WHERE id = ?""",
-                (rec_id,),
-            ).fetchone()
-        finally:
-            conn.close()
-
-        log.warning(
-            "Kill switch v2: operator ignored recomendación id=%d", rec_id,
-        )
-    except HTTPException:
-        raise
-    except Exception as e:
-        log.error(
-            "POST /kill_switch/recommendations/%d/ignore failed: %s",
-            rec_id, e, exc_info=True,
-        )
-        raise HTTPException(
-            status_code=500,
-            detail=f"ignore failed: {type(e).__name__}: {e}",
-        )
-
-    return {
-        "id": updated[0], "ts": updated[1], "triggered_by": updated[2],
-        "slider_value": updated[3], "projected_pnl": updated[4],
-        "projected_dd": updated[5], "status": updated[6],
-        "applied_ts": updated[7], "applied_by": updated[8],
-    }
-
-
 @app.get("/webhook/test", summary="Probar webhook y Telegram directo", dependencies=[Depends(verify_api_key)])
 def test_webhook():
     cfg     = load_config()
@@ -909,309 +628,7 @@ def test_webhook():
     return {"ok": overall_ok, **results}
 
 
-# ── Auto-Tune Endpoints ──────────────────────────────────────────────────────
-
-@app.get("/tune/latest", summary="Latest tune result")
-def tune_latest():
-    """Returns the most recent tune_result row (with parsed results_json) or null."""
-    con = get_db()
-    row = con.execute(
-        "SELECT * FROM tune_results ORDER BY id DESC LIMIT 1"
-    ).fetchone()
-    con.close()
-    if not row:
-        return None
-    result = dict(row)
-    # Parse results_json so the frontend gets an object, not a string
-    if result.get("results_json"):
-        try:
-            result["results_json"] = json.loads(result["results_json"])
-        except (json.JSONDecodeError, TypeError):
-            pass
-    return result
-
-
-@app.post("/tune/apply", summary="Apply pending tune proposal",
-          dependencies=[Depends(verify_api_key)])
-def tune_apply():
-    """Applies the latest pending tune proposal to config.json symbol_overrides."""
-    con = get_db()
-    row = con.execute(
-        "SELECT * FROM tune_results WHERE status = 'pending' ORDER BY id DESC LIMIT 1"
-    ).fetchone()
-    if not row:
-        con.close()
-        raise HTTPException(status_code=404, detail="No pending tune proposal found")
-
-    result = dict(row)
-    tune_id = result["id"]
-
-    # Parse results_json to extract CHANGE recommendations
-    try:
-        results = json.loads(result["results_json"]) if result.get("results_json") else {}
-    except (json.JSONDecodeError, TypeError):
-        con.close()
-        raise HTTPException(status_code=500, detail="Invalid results_json in tune proposal")
-
-    # Create config backup
-    now_str = datetime.now(timezone.utc).strftime("%Y%m%d_%H%M%S")
-    backup_name = f"config_backup_{now_str}.json"
-    backup_path = os.path.join(SCRIPT_DIR, backup_name)
-    if os.path.exists(CONFIG_FILE):
-        shutil.copy2(CONFIG_FILE, backup_path)
-
-    # Load config and apply changes
-    cfg = load_config()
-    overrides = cfg.get("symbol_overrides", {})
-    applied_count = 0
-
-    # Extract CHANGE recommendations from results
-    recommendations = results.get("recommendations", [])
-    for rec in recommendations:
-        if rec.get("action") != "CHANGE":
-            continue
-        symbol = rec.get("symbol", "")
-        params = rec.get("params", {})
-        if symbol and params:
-            if symbol not in overrides:
-                overrides[symbol] = {}
-            overrides[symbol].update(params)
-            applied_count += 1
-
-    cfg["symbol_overrides"] = overrides
-    with open(CONFIG_FILE, "w", encoding="utf-8") as f:
-        json.dump(cfg, f, indent=2, ensure_ascii=False)
-    log.info(f"Auto-tune applied: {applied_count} changes, backup: {backup_name}")
-
-    # Update tune_result status
-    con.execute(
-        "UPDATE tune_results SET status = 'applied', applied_ts = ?, changes_count = ? WHERE id = ?",
-        (datetime.now(timezone.utc).isoformat(), applied_count, tune_id)
-    )
-    con.commit()
-    con.close()
-
-    return {"ok": True, "applied": applied_count, "backup": backup_name}
-
-
-@app.post("/tune/reject", summary="Reject pending tune proposal",
-          dependencies=[Depends(verify_api_key)])
-def tune_reject():
-    """Rejects the latest pending tune proposal."""
-    con = get_db()
-    row = con.execute(
-        "SELECT id FROM tune_results WHERE status = 'pending' ORDER BY id DESC LIMIT 1"
-    ).fetchone()
-    if not row:
-        con.close()
-        raise HTTPException(status_code=404, detail="No pending tune proposal found")
-
-    con.execute(
-        "UPDATE tune_results SET status = 'rejected' WHERE id = ?",
-        (row["id"],)
-    )
-    con.commit()
-    con.close()
-    return {"ok": True}
-
-
-# ── Kill switch / health endpoints (#138) ─────────────────────────────
-
-
-class ReactivateRequest(BaseModel):
-    reason: str = "manual"
-
-
-@app.get("/health/symbols", dependencies=[Depends(verify_api_key)])
-def get_health_symbols():
-    """List current health state per symbol."""
-    con = get_db()
-    try:
-        rows = con.execute(
-            """SELECT symbol, state, state_since, last_evaluated_at,
-                      last_metrics_json, manual_override
-               FROM symbol_health
-               ORDER BY symbol"""
-        ).fetchall()
-    finally:
-        con.close()
-    cols = ("symbol", "state", "state_since", "last_evaluated_at",
-            "last_metrics_json", "manual_override")
-    return {"symbols": [dict(zip(cols, r)) for r in rows]}
-
-
-@app.get("/health/events", dependencies=[Depends(verify_api_key)])
-def get_health_events(
-    symbol: Optional[str] = None,
-    limit: int = Query(50, ge=1, le=500, description="Max rows to return (capped to prevent unbounded scans)"),
-):
-    """Transition history. Optionally filter by symbol."""
-    con = get_db()
-    try:
-        if symbol:
-            rows = con.execute(
-                """SELECT id, symbol, from_state, to_state, trigger_reason,
-                          metrics_json, ts
-                   FROM symbol_health_events WHERE symbol=?
-                   ORDER BY ts DESC LIMIT ?""",
-                (symbol, limit),
-            ).fetchall()
-        else:
-            rows = con.execute(
-                """SELECT id, symbol, from_state, to_state, trigger_reason,
-                          metrics_json, ts
-                   FROM symbol_health_events ORDER BY ts DESC LIMIT ?""",
-                (limit,),
-            ).fetchall()
-    finally:
-        con.close()
-    cols = ("id", "symbol", "from_state", "to_state", "trigger_reason",
-            "metrics_json", "ts")
-    return {"events": [dict(zip(cols, r)) for r in rows]}
-
-
-# ── Notification center endpoints (#162 PR C) ────────────────────────
-@app.get("/notifications", dependencies=[Depends(verify_api_key)])
-def get_notifications(
-    unread: bool = True,
-    limit: int = Query(50, ge=1, le=200,
-                        description="Max rows returned (capped to prevent unbounded scans)"),
-):
-    """List notifications recorded by the notifier.
-
-    By default returns only unread entries; pass ?unread=false to include
-    read ones too. Sorted most-recent-first.
-    """
-    from notifier._storage import list_unread
-    if not unread:
-        # Full list (both read + unread) — use a direct query since list_unread
-        # filters on read_at IS NULL.
-        con = get_db()
-        try:
-            rows = con.execute(
-                """SELECT id, event_type, event_key, priority, payload_json,
-                          channels_sent, delivery_status, sent_at, read_at, error_log
-                   FROM notifications_sent
-                   ORDER BY sent_at DESC
-                   LIMIT ?""",
-                (limit,),
-            ).fetchall()
-        finally:
-            con.close()
-        cols = ("id", "event_type", "event_key", "priority", "payload_json",
-                "channels_sent", "delivery_status", "sent_at", "read_at", "error_log")
-        return {"notifications": [dict(zip(cols, r)) for r in rows]}
-    return {"notifications": list_unread(limit=limit)}
-
-
-@app.post("/notifications/{notif_id}/read", dependencies=[Depends(verify_api_key)])
-def post_notification_read(notif_id: int):
-    """Mark a single notification as read."""
-    from notifier._storage import mark_read
-    mark_read(notif_id)
-    return {"ok": True, "id": notif_id}
-
-
-@app.post("/notifications/read-all", dependencies=[Depends(verify_api_key)])
-def post_notifications_read_all():
-    """Mark all currently-unread notifications as read. Returns how many were updated."""
-    from notifier._storage import mark_all_read
-    n = mark_all_read()
-    return {"ok": True, "marked": n}
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-#  KILL SWITCH OBSERVABILITY ENDPOINTS (#187 phase 1)
-# ─────────────────────────────────────────────────────────────────────────────
-
-@app.get("/kill_switch/decisions", dependencies=[Depends(verify_api_key)])
-def get_kill_switch_decisions(
-    symbol: Optional[str] = None,
-    engine: Optional[str] = None,
-    since: Optional[str] = None,
-    limit: int = Query(50, ge=1, le=200),
-):
-    """Kill switch decision log (#187 phase 1). Filter by symbol/engine/since ts."""
-    import observability
-    rows = observability.query_decisions(
-        symbol=symbol, engine=engine, since=since, limit=limit,
-    )
-    return {"decisions": rows}
-
-
-@app.get("/kill_switch/current_state", dependencies=[Depends(verify_api_key)])
-def get_kill_switch_current_state(engine: str = "v1"):
-    """Current tier state per symbol + portfolio aggregate (#187 phase 1)."""
-    import observability
-    return observability.get_current_state(engine=engine)
-
-
-@app.get("/health/dashboard", dependencies=[Depends(verify_api_key)])
-def get_health_dashboard():
-    """B6: single-shot consolidated state for the kill switch dashboard.
-
-    Returns per-symbol full state + portfolio aggregate + 24h alert summary.
-    Read-only; safe even when kill_switch.enabled=False (returns last-evaluated
-    snapshot).
-    """
-    from health import get_dashboard_state
-    cfg = load_config()
-    return get_dashboard_state(cfg)
-
-
-@app.post("/health/reactivate/{symbol}", dependencies=[Depends(verify_api_key)])
-def post_health_reactivate(symbol: str, body: ReactivateRequest):
-    """Manually reactivate a PAUSED symbol — transitions PAUSED → PROBATION (B5 #199)."""
-    from health import reactivate_symbol, get_symbol_state
-    cfg = load_config()
-    reactivate_symbol(symbol.upper(), reason=body.reason, cfg=cfg)
-    return {"ok": True, "symbol": symbol.upper(), "state": get_symbol_state(symbol.upper())}
-
-
-@app.get("/health", summary="Health check for monitoring and Docker")
-def health_check():
-    """Returns system health status. HTTP 200 = healthy, 503 = degraded."""
-    checks = {}
-
-    # Database connectivity
-    try:
-        con = get_db()
-        con.execute("SELECT 1")
-        con.close()
-        checks["database"] = "ok"
-    except Exception as e:
-        checks["database"] = f"error: {e}"
-
-    # Scanner thread status
-    checks["scanner"] = "ok" if _scanner_state.get("running") else "stopped"
-
-    # Last scan freshness
-    last_ts = _scanner_state.get("last_scan_ts")
-    if last_ts:
-        try:
-            last_dt = datetime.fromisoformat(last_ts)
-            age_sec = (datetime.now(timezone.utc) - last_dt).total_seconds()
-            cfg = load_config()
-            interval = cfg.get("scan_interval_sec", 300)
-            checks["scan_freshness"] = "ok" if age_sec < interval * 3 else f"stale ({int(age_sec)}s ago)"
-        except Exception:
-            checks["scan_freshness"] = "unknown"
-    else:
-        checks["scan_freshness"] = "no_scans_yet"
-
-    # Stats
-    checks["scans_total"] = _scanner_state.get("scans_total", 0)
-    checks["signals_total"] = _scanner_state.get("signals_total", 0)
-    checks["errors"] = _scanner_state.get("errors", 0)
-
-    healthy = checks["database"] == "ok" and checks["scanner"] == "ok"
-    status_code = 200 if healthy else 503
-
-    from fastapi.responses import JSONResponse
-    return JSONResponse(
-        content={"healthy": healthy, "checks": checks},
-        status_code=status_code
-    )
+# ── Auto-Tune, Kill switch, Health, Notifications → moved to api/ in PR6 ──────
 
 
 # ─────────────────────────────────────────────────────────────────────────────
